+++ packages/plugins/type-distributor/src/headless/import/index.ts
@@ -1,0 +1,1 @@
+export { loadFromLocal } from './load-file.helper'