+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the monorepo and its configuration including the workflows will be documented here
 
+## 2025-12-18
+### Changed
+- Updated workspace scripts and dependency versions related to the type-distributor plugin
+
 ## 2025-12-10
 ### Changed
 - Use the scripts defined in the `package.json` for the workflow so that we can build communication and network element for the autodoc